03.02.2026 ui rework

+++ Assets/Scripts/SaveS/LevelSaveData.cs
@@ -6,10 +6,10 @@ public struct LevelSaveData
     public LevelID Level;
     public int Score;
     public int Stars;
-    public int Time;
+    public float Time;
     public bool IsComplite;
 
-    public LevelSaveData(LevelID level, int score, int stars, int time)
+    public LevelSaveData(LevelID level, int score, int stars, float time)
     {
         Level = level;
         Score = score;
